GenreWordReplacerを個別モジュールに分離し、独立性を担保
